move to logic folder
